Moved sub group implementation to runtime library.
