+++ include/neso_particles/particle_sub_group/particle_sub_group.hpp
@@ -9,10 +9,8 @@
 #include "particle_group_partition.hpp"
 #include "particle_loop_sub_group.hpp"
 #include "particle_sub_group_base.hpp"
-#include "particle_sub_group_base_impl.hpp"
 #include "sub_group_selector.hpp"
 #include "sub_group_selector_base.hpp"
-#include "sub_group_selector_base_impl.hpp"
 
 namespace NESO::Particles {
